gemma2b/requirements.txt
pandas
Faker
chromadb
numpy
matplotlib
seaborn
torch --index-url https://download.pytorch.org/whl/cpu
transformers
accelerate
peft
datasets 
bitsandbytes
python-docx 
PyMuPDF 
tqdm
pywin32